Some refactoring, delete apk
