+++ petkeeper/src/main/java/com/projet/petkeeper/chat/ChatRootScreen.kt
@@ -1,17 +1,23 @@
 package com.projet.petkeeper.chat
 
 import android.util.Log
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material3.CenterAlignedTopAppBar
 import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
@@ -20,17 +26,18 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import coil.compose.rememberAsyncImagePainter
 import com.projet.petkeeper.data.UserData
 import com.projet.petkeeper.data.UserPair
 import com.projet.petkeeper.data.userSamples
 import com.projet.petkeeper.ui.PetKeeperUIState
 import com.projet.petkeeper.ui.theme.PetkeeperTheme
-import com.projet.petkeeper.utils.UserProfileImageIcon
 import kotlinx.coroutines.runBlocking
 
 
@@ -134,6 +141,30 @@ fun ChatCard(
     }
 
 }
+
+
+@Composable
+fun UserProfileImageIcon(userData: UserData?) {
+    Surface(
+        modifier = Modifier.size(40.dp),
+        shape = CircleShape,
+        color = Color.Transparent,
+        border = BorderStroke(
+            1.5.dp,
+            MaterialTheme.colorScheme.secondary
+        )
+    ) {
+        val profileImageUrl = userData?.photoURL // Get the profile image URL
+        val painter = rememberAsyncImagePainter(model = profileImageUrl)
+        Image(
+            painter = painter,
+            contentDescription = "User Profile",
+            modifier = Modifier.fillMaxSize()
+        )
+    }
+}
+
+
 /**
  * Preview function to display a preview of the ChatRootScreen composable.
  */
